+++ packages/expo-modules-core/CHANGELOG.md
@@ -27,6 +27,7 @@
 ### 💡 Others
 
 - [iOS] Removed some runtime type checks for dynamic types. ([#40611](https://github.com/expo/expo/pull/40611) by [@tsapeta](https://github.com/tsapeta))
+- [iOS] Added base props support for SwiftUI integration. ([#40492](https://github.com/expo/expo/pull/40492) by [@kudo](https://github.com/kudo))
 
 ## 3.0.22 — 2025-10-20